+++ src/bindpy/bindings.cpp
@@ -12,7 +12,7 @@ namespace py = pybind11;
 
 PYBIND11_MODULE(_astroflow_core, m) {
 
-  bind_dedispersed_data<uint32_t>(m, "DedisperedData");
+  bind_dedispersed_data<uint64_t>(m, "DedisperedData");
   bind_filterbank(m);
   bind_spectrum<uint8_t>(m, "Spectrum8");
   bind_spectrum<uint16_t>(m, "Spectrum16");
